+++ newton/_src/geometry/__init__.py
@@ -18,6 +18,7 @@ from .collision_primitive import (
     collide_sphere_cylinder,
     collide_sphere_sphere,
 )
+from .contact_match import MATCH_BROKEN, MATCH_NOT_FOUND
 from .flags import ParticleFlags, ShapeFlags
 from .inertia import compute_inertia_shape, compute_inertia_sphere, transform_inertia
 from .sdf_utils import SDF
@@ -32,6 +33,8 @@ from .types import (
 from .utils import compute_shape_radius
 
 __all__ = [
+    "MATCH_BROKEN",
+    "MATCH_NOT_FOUND",
     "SDF",
     "BroadPhaseAllPairs",
     "BroadPhaseExplicit",